media: i2c: ov9282: Fix reset GPIO polarity
The ov9282 sensor has a reset line that is active low not active
high. Currently the logic for this is inverted.

The right way to define the reset line is to declare it active low in the
DTS and invert the logic currently contained in the driver.

This fix is a clone of a similar fix for the imx412. It appears to be a
common code-base which has shared this same polarity bug.

Commit: bb25f071fc92 ("media: i2c: imx412: Fix reset GPIO polarity")

We need to apply precisely that same fix to this driver too.

Fixes: 14ea315bbeb7 ("media: i2c: Add ov9282 camera sensor driver")
Cc: stable@vger.kernel.org
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>