media: camss: Add indirection wrapper to get CSID RX timestamp
Some versions of CAMSS hardware are capable of timestamping the SOF and EOF
CSI bus protocol states in a CSID register. The timer used for this value
can be aligned to a kernel CLOCK_BOOTTIME counter.

Provide an indirection to the CSID driver to allow the CSID driver to
return a ktime_get_ns() value. We will subsequently allow some CSID drivers
to return an SOF aligned BOOTTIME value when it can be supported.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>