media: camss: Add a routine to return a CAMSS Qtimer value in nanoseconds
Qtimer is a free-running timer value that does not halt in system suspend.
The Qtimer values are specific to a given Virtual Channel - VC - stream so,
we need to offset based on the passed VC index.

In upstream CAMSS we use a fixed mapping for the Raw Data Interface - RDI
to VCs, hence VC0 => RDI0, VC1 => RDI1.

We convert from the returned Qtimer value to nanoseconds so that the
calling routine doesn't have to know or care what the clock-rate of the
Qtimer is. In the case of CSID-Gen2, that Qtimer rate is 19.2MHz.

For the RDI Qtimer values are available for CURRENT_SOF, CURRENT_EOF and
PREV_SOF, PREV_EOF with SOF and EOF representing the MIPI CSI bus-states
Start-Of-Frame and End-Of-Frame.

The reference Qualcomm Camx code only interacts with CURRENT_SOF so this
implementation follows suit.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>