media: camss: Enable Qtimer based V4L2 buffers
From sdm845 onwards CAMSS ships with Qtimer registers embedded in various
parts of the CAMSS IP blocks. The Qtimer registers are free-running timers
which do not stop in suspend.

The Qualcomm camx stack makes use of these timers by

1. Taking an initial snapshot of the get_monotonic_boottime64()
2. On every write-master complete taking a snapshot of the Qtimer value
   associated with CSI SOF.
3. Appending the difference between the current CSI SOF Qtimer and the
   previous frame CSI SOF Qtimer to the initial boottime.

In this way the free-running Qtimer is appended to a fixed boottime value.

The main advantage to this method of timestamping is we can use the
hardware stuffed CSI SOF Qtimer value, which is latched on the falling
clock of the CSI SOF bus state.

The main disadvantage is the initial latency snapshotting boottime64()
above introduces a permanent skew into the delivered values.

A similar algorithm has been implemented here as given above, though
simpler than in the Qualcomm original in that we don't convert to a
timespec inside of the driver.

The new flag V4L2_BUF_FLAG_TIMESTAMP_BOOTTIME informs user-space to do a
clock_gettime(CLOCK_BOOTTIME, &ts); to resolve the given buffer timestamp
to a timespec.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>