arm64: dts: qcom: x1e80100: Add MIPI CSI PHY nodes
Add csiphy nodes for

- csiphy0
- csiphy1
- csiphy2
- csiphy4

The irregular naming of the PHYs comes directly from the hardware which for
whatever reason skipped csiphy3.

Separating the nodes from CAMSS as we have done with the sensor I2C bus aka
the CCI interface is justified since the CSIPHYs have their own pinouts and
voltage rails.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>

827878421ff03a954c9a7f6904290c5c51b72064

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>